requirements.txt
torch==2.4.0
xformers==0.0.27.post2
diffusers==0.32.2
transformers==4.44.2
accelerate
peft
safetensors
huggingface_hub
Pillow
open_clip_torch
pandas
